Move map to its own file
